requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
langchain==1.0.7  # Already installed; keeps 1.0 features
langchain-community==0.3.31  # Use a 0.3.x version compatible with langchain 1.0.7 (avoids potential mismatches; latest 0.3 is 0.3.31)
langchain-ollama==1.0.0  # Latest; works with 1.0
langchain-core==1.0.5  # Latest; required by langchain 1.0.7
python-dotenv==1.0.1
requests==2.32.5  # Matches langchain-community's minimum; includes all prior security patches
faiss-cpu==1.12.0  # Latest compatible version
sentence-transformers==3.1.1  # Updated for better embeddings
torch==2.9.1  # Updated to latest stable (CPU-compatible)
pytest==8.3.3  # For tests